% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stat_models.R
\name{calculate_stat_models}
\alias{calculate_stat_models}
\title{Calculate function for set of models}
\usage{
calculate_stat_models(dataset, models, fun, ...)
}
\arguments{
\item{dataset}{Dataset to operate on.}

\item{models}{Models to calculate for.}

\item{fun}{Function to calculate (i.e. t.test).}

\item{...}{Parameters passed to \code{fun}.}
}
\value{
Data frame with results of function for each model.
}
\description{
Calculate function for set of models
}
\examples{
library(ggplot2)
library(dplyr)
data(diamonds)
diamonds <- diamonds \%>\%
  filter(color \%in\% c("D", "E"))
models <- list(
  carat ~ color,
  depth ~ color,
  x ~ color
)
calculate_stat_models(diamonds, models, t.test)
}
