% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_function_in_package.R
\name{find_function_in_package}
\alias{find_function_in_package}
\title{Find function in package}
\usage{
find_function_in_package(f)
}
\arguments{
\item{f}{function or character with name of the function}
}
\value{
character name of the package that contains the function
}
\description{
Find function in package
}
\examples{
package_name <- find_function_in_package(t.test)
package_name <- find_function_in_package("t.test")

}
