% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statHelper-package.R
\docType{package}
\name{statHelper-package}
\alias{statHelper}
\alias{statHelper-package}
\title{statHelper: Functions to Help with Statistical Evaluation of Data}
\description{
Functions to help with mainly statistical evaluation of data.
}
\author{
\strong{Maintainer}: Jan Caha \email{jan.caha@outlook.com} (\href{https://orcid.org/0000-0003-0165-0606}{ORCID})

}
\keyword{internal}
