% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_frequency.R
\name{stats_frequency}
\alias{stats_frequency}
\title{Calculates frequency statistics of data}
\usage{
stats_frequency(data, ..., .select_function = is.factor)
}
\arguments{
\item{data}{A data frame to process.}

\item{...}{Grouping variables to be applied before calculation.}

\item{.select_function}{Function used to select data columns for which the stats will be calculated. Default value is \code{is.factor}.}
}
\value{
Data frame containing stats.
}
\description{
Calculates frequency statistics of data
}
\examples{
library(ggplot2)
data("diamonds")
ds_stats <- stats_frequency(diamonds)
ds_stats <- stats_frequency(diamonds, color)
}
