% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_numeric.R
\name{stats_numeric}
\alias{stats_numeric}
\title{Calculates numeric statistics of data}
\usage{
stats_numeric(
  data,
  ...,
  stats_to_calculate = NULL,
  .select_function = is.numeric
)
}
\arguments{
\item{data}{A data frame to process.}

\item{...}{Grouping variables to be applied before calculation.}

\item{stats_to_calculate}{List of statistics function to calculate. Default value is \code{NULL} which means that standard functions (n, mean, sd, median, iqr, min and max) are calculate.}

\item{.select_function}{Function used to select data columns for which the stats will be calculated. Default value is \code{is.numeric}.}
}
\value{
Data frame containing stats.
}
\description{
Calculates numeric statistics of data
}
\examples{
library(ggplot2)
library(dplyr)
data("diamonds")
ds_stats <- stats_numeric(diamonds)
stats_to_calculate <- list(n = ~ n(), mean = ~ mean(.x, na.rm = TRUE), sd = ~ sd(.x, na.rm = TRUE))
ds_stats <- stats_numeric(diamonds, stats_to_calculate = stats_to_calculate)

}
