% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalitas.R
\name{uji_normalitas}
\alias{uji_normalitas}
\title{Uji Normalitas untuk Kolom Numerik}
\usage{
uji_normalitas(data, columns = NULL, methods = c("shapiro", "ks", "ad"))
}
\arguments{
\item{data}{Data frame yang berisi data yang akan diuji normalitasnya.}

\item{columns}{Vektor nama kolom yang ingin diuji. Jika `NULL`, semua kolom numerik akan diuji.}

\item{methods}{Vektor metode uji normalitas yang akan digunakan. Defaultnya adalah `c("shapiro", "ks", "ad")`.
Metode yang dapat dipilih adalah:
- "shapiro" untuk uji Shapiro-Wilk
- "ks" untuk uji Kolmogorov-Smirnov
- "ad" untuk uji Anderson-Darling.}
}
\value{
Data frame dengan hasil uji normalitas, yang mencakup:
  - `Column`: Nama kolom yang diuji
  - `Method`: Nama metode uji yang digunakan
  - `Statistic`: Nilai statistik uji
  - `P.Value`: Nilai p dari uji normalitas
}
\description{
Fungsi ini melakukan uji normalitas untuk kolom numerik dalam dataset menggunakan beberapa metode uji,
termasuk Shapiro-Wilk, Kolmogorov-Smirnov, dan Anderson-Darling. Fungsi ini mengembalikan hasil statistik uji
dan p-value untuk setiap kolom yang diuji.
}
\details{
Fungsi ini menggunakan paket `nortest` untuk uji Anderson-Darling, serta fungsi `shapiro.test` dan `ks.test` dari R
untuk uji Shapiro-Wilk dan Kolmogorov-Smirnov. Fungsi ini mengembalikan hasil statistik uji dan p-value untuk
setiap kolom numerik dalam data frame.
}
\examples{
# Membuat data contoh
data <- data.frame(a = rnorm(100), b = rnorm(100), c = runif(100))

# Uji normalitas untuk kolom 'a' dan 'b' dengan metode Shapiro-Wilk dan Anderson-Darling
uji_normalitas(data, columns = c("a", "b"), methods = c("shapiro", "ad"))

# Uji normalitas untuk semua kolom dengan semua metode
uji_normalitas(data)

}
