% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Compute z-scores}
\usage{
z_score(x, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na_rm}{Logical. Should missing values be removed before calculation?}
}
\value{
A numeric vector of z-scores.
}
\description{
Standardizes a numeric vector into z-scores.
}
\examples{
z_score(c(1, 2, 3, 4, 5))
}
