% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{z_score}
\alias{z_score}
\title{Calculate z-score}
\usage{
z_score(x, mean = NULL, sd = NULL)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{mean}{The mean to use for standardization. If NULL, the mean of x is used.}

\item{sd}{The standard deviation to use for standardization. If NULL, the sd of x is used.}
}
\value{
A numeric vector of z-scores.
}
\description{
This function calculates the z-score (standard score) for a value or vector of values.
}
\examples{
z_score(c(1, 2, 3, 4, 5))
z_score(c(10, 20, 30), mean = 20, sd = 10)
}
